dense_retrieval/requirements.txt
pandas
sentence-transformers
faiss-cpu
numpy
